Pause display link when app goes to background
I observed on a real iPad Air (5th gen, iPadOS 26.1) that enterFrame() was still being called when
the app went to the background. This caused repeated GPU warnings:

---
IOGPUMetalError: Insufficient Permission (to submit GPU work from background) (00000006:kIOGPUCommandBufferCallbackErrorBackgroundExecutionNotPermitted)
---
Execution of the command buffer was aborted due to an error during execution. Insufficient Permission (to submit GPU work from background) (00000006:kIOGPUCommandBufferCallbackErrorBackgroundExecutionNotPermitted)
---

The warnings fill the console, and if you scroll up in Xcode, memory usage spikes and can crash the
app. CPU usage while in the background was around 30-50% before this fix and now drops to 0.
I did not see this on the simulator (iPhone 17 Pro, iOS 26.1) and haven't tested other devices.

I fixed it by pausing the displayLink when the app goes to the background and resuming it when the
app comes to the foreground.